requirements.txt
# Core ML Libraries
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
openai-whisper>=20231117
pyannote.audio>=3.1.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.1
pyaudio>=0.2.11
webrtcvad>=2.0.10

# Web Framework
fastapi>=0.100.0
websockets>=11.0.2
uvicorn>=0.22.0
python-multipart>=0.0.6

# Database
pymongo>=4.3.3
motor>=3.2.0

# Utilities
numpy>=1.24.3
scipy>=1.11.1
scikit-learn>=1.3.0
pandas>=2.0.3
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0

# Export formats
jinja2>=3.1.2
reportlab>=4.0.4

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
